Move make_logs_queue in utils/logs.py
