+++ src/aleph/vm/controllers/qemu/instance.py
@@ -5,14 +5,15 @@ import shutil
 from asyncio import Task
 from asyncio.subprocess import Process
 from pathlib import Path
-from typing import Callable, Generic, Optional, TypedDict, TypeVar, Union
+from typing import Callable, Generic, Optional, TypeVar, Union
 
 import psutil
+
+from aleph.vm.utils.logs import make_logs_queue
 from aleph_message.models import ItemHash
 from aleph_message.models.execution.environment import MachineResources
 from aleph_message.models.execution.instance import RootfsVolume
 from aleph_message.models.execution.volume import PersistentVolume, VolumePersistence
-from systemd import journal
 
 from aleph.vm.conf import settings
 from aleph.vm.controllers.configuration import (
@@ -84,57 +85,6 @@ class AlephQemuResources(AlephFirecrackerResources):
 
 
 ConfigurationType = TypeVar("ConfigurationType")
-
-
-class EntryDict(TypedDict):
-    SYSLOG_IDENTIFIER: str
-    MESSAGE: str
-
-
-def make_logs_queue(stdout_identifier, stderr_identifier, skip_past=False) -> tuple[asyncio.Queue, Callable[[], None]]:
-    """Create a queue which streams the logs for the process.
-
-    @param stdout_identifier: journald identifier for process stdout
-    @param stderr_identifier: journald identifier for process stderr
-    @param skip_past: Skip past history.
-    @return: queue and function to cancel the queue.
-
-    The consumer is required to call the queue cancel function when it's done consuming the queue.
-
-    Works by creating a journald reader, and using `add_reader` to call a callback when
-    data is available for reading.
-    In the callback we check the message type and fill the queue accordingly
-
-    For more information refer to the sd-journal(3) manpage
-    and systemd.journal module documentation.
-    """
-    r = journal.Reader()
-    r.add_match(SYSLOG_IDENTIFIER=stdout_identifier)
-    r.add_match(SYSLOG_IDENTIFIER=stderr_identifier)
-    queue: asyncio.Queue = asyncio.Queue(maxsize=1000)
-
-    def _ready_for_read() -> None:
-        change_type = r.process()  # reset fd status
-        if change_type != journal.APPEND:
-            return
-        entry: EntryDict
-        for entry in r:
-            log_type = "stdout" if entry["SYSLOG_IDENTIFIER"] == stdout_identifier else "stderr"
-            msg = entry["MESSAGE"]
-            asyncio.create_task(queue.put((log_type, msg)))
-
-    if skip_past:
-        r.seek_tail()
-
-    loop = asyncio.get_event_loop()
-    loop.add_reader(r.fileno(), _ready_for_read)
-
-    def do_cancel():
-        logger.info(f"cancelling reader {r}")
-        loop.remove_reader(r.fileno())
-        r.close()
-
-    return queue, do_cancel
 
 
 class AlephQemuInstance(Generic[ConfigurationType], CloudInitMixin, AlephVmControllerInterface):
